pyproject.toml
[tool.poetry]
name = "RAG-model-practice"
version = "0.1.0"
description = "A RAG implementation in Python"
authors = ["Igor Asmykovich <igor.asmykovich@gmail.com>"]
license = "Apache 2.0"
packages = [{ include = "rag", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = "~0.111.0"
uvicorn = { extras = ["standard"], version = "~0.29.0" }
pydantic = "~2.7.0"
langchain = "~0.2.0"
pinecone-client = "~4.1.0"
boto3 = "~1.38.34"
pymupdf = "~1.23.0"
pyyaml = "~6.0.1"
langchain-community = "^0.2.0"
sentence-transformers = "^2.7.0"

[tool.poetry.scripts]
rag-run = "rag.mainRAG:main"

[poetry.group.dev.dependencies]
pytest = "~8.2.0"
ruff = "~0.4.0"
black = "~24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"